Made `node_index_in_block` method return `u32` (#814)
* refactor: make `node_index_in_block` method return `u32`

* docs: update CHANGELOG.md

* refactor: rename constants

* docs: update doc comment

* docs: fix docs building

+++ objects/src/notes/location.rs
@@ -24,10 +24,10 @@ impl NoteLocation {
     ///
     /// # Note
     ///
-    /// The height of the Merkle tree is `BLOCK_OUTPUT_NOTES_TREE_DEPTH`. Thus, the maximum index
-    /// is `2 ^ BLOCK_OUTPUT_NOTES_TREE_DEPTH - 1`.
-    pub fn node_index_in_block(&self) -> u64 {
-        self.node_index_in_block as u64
+    /// The height of the Merkle tree is [crate::constants::BLOCK_NOTES_TREE_DEPTH].
+    /// Thus, the maximum index is `2 ^ BLOCK_NOTES_TREE_DEPTH - 1`.
+    pub fn node_index_in_block(&self) -> u32 {
+        self.node_index_in_block
     }
 }
